Moved SearchBar and NameCollisionDialog vues to ./search subfolder.
