+++ src/components/search/NameCollisionDialog.vue
