+++ src/pages/MobileSearch.vue
@@ -49,7 +49,7 @@ import {computed, defineComponent, inject, onBeforeUnmount, onMounted, ref, watc
 import {useRoute} from "vue-router";
 import router from "@/router";
 import {MEDIUM_BREAKPOINT} from "@/App.vue";
-import SearchBar from "@/components/SearchBar.vue";
+import SearchBar from "@/components/search/SearchBar.vue";
 import Footer from "@/components/Footer.vue";
 
 export default defineComponent({
